% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-formatters.R
\name{md_formatters}
\alias{md_formatters}
\alias{md_bold}
\alias{md_italic}
\alias{md_strikethrough}
\alias{md_underline}
\alias{md_spoiler}
\title{Markdown formatters}
\usage{
md_bold(string)

md_italic(string)

md_strikethrough(string)

md_underline(string)

md_spoiler(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
A character vector with the same length as the input.
}
\description{
Format a character vector with Markdown.
}
\details{
The Markdown formatters in the \href{https://k5cents.github.io/gluedown/}{gluedown} package use
\href{https://glue.tidyverse.org/}{glue}, which support strings and not character vectors. For
consistency, since all string-altering functions in this package support and return character
vectors, vectorised Markdown-formatters are also included. The Markdown syntax used here is
tailored for Discord.
}
