% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menace-package.R
\docType{package}
\name{menace-package}
\alias{menace}
\alias{menace-package}
\title{menace: Unleash your inner menace}
\description{
Tools to make you a menace.
}
\author{
\strong{Maintainer}: Adam Shen \email{adamshen1@gmail.com}

}
\keyword{internal}
