% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{mock}
\alias{mock}
\title{Return a character vector mockingly}
\usage{
mock(string, method = "alternate")
}
\arguments{
\item{string}{A character vector.}

\item{method}{A string denoting the method to use: either \code{"alternate"} (default) or \code{"random"}.}
}
\value{
A character vector with the same length as the input.
}
\description{
Return a character vector mockingly by altering the case of letters. See \emph{Details}.
}
\details{
When \code{method = "alternate"}, the first letter of each string is lowercased. Letters then
alternate between lowercase and uppercase. When \code{method = "random"}, 60\% of the letters
(at random) in each string will swap to the other case.

When the output is to be pasted to Microsoft Teams, mocked messages can be formatted using the
included \link[=teams_md]{teams formatters} and copied to the clipboard using \code{clipr::write_clip()}.
For other destinations, the \href{https://k5cents.github.io/gluedown/}{gluedown} package provides
GitHub Flavoured Markdown formatters.
}
\examples{
library(magrittr)

# Single string
msg <- "Delete my number from your phone."
mock(msg)
mock(msg, "random")

# Write to clipboard and paste to Teams
# mock(msg) \%>\%
#   teams_italic() \%>\%
#   clipr::write_clip()

# Character vector
msgs <- c("OMG, you're sooooo annoying", "Don't ever speak to me again")
mock(msgs)
mock(msgs, "random")

# Write to clipboard and paste to Teams
# mock(msgs) \%>\%
#   teams_bold() \%>\%
#   clipr::write_clip()
}
