% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shout.R
\name{shout}
\alias{shout}
\title{Return a character vector loudly}
\usage{
shout(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
A character vector with the same length as the input.
}
\description{
Return a character vector loudly by uppercasing all letters and increasing spacing. Best used
with a bold-formatting.
}
\details{
Inspired by Ryan Kavanagh.
}
\examples{
# example code
msg <- "listen up, you little jerk"
shout(msg)

# Recommended to use with bold-formatting.
# Write to clipboard and paste elsewhere
# msg |>
#   shout() |>
#   clipr::write_clip()
}
