% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams_md.R
\name{teams_md}
\alias{teams_md}
\alias{teams_bold}
\alias{teams_italic}
\alias{teams_strikethrough}
\title{Markdown formatting for Microsoft Teams}
\usage{
teams_bold(string)

teams_italic(string)

teams_strikethrough(string)
}
\arguments{
\item{string}{A character vector.}
}
\description{
Format a character vector for Microsoft Teams.
}
\details{
Microsoft Teams uses a different Markdown syntax. See \href{https://support.microsoft.com/en-us/office/use-markdown-formatting-in-microsoft-teams-4d10bd65-55e2-4b2d-a1f3-2bebdcd2c772}{here}
for a full list of formats.
}
