% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menace-package.R
\docType{package}
\name{menace-package}
\alias{menace}
\alias{menace-package}
\title{menace: Unleash Your Inner Menace}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Text-formatting tools to help you unleash your inner menace.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adamoshen.github.io/menace}
  \item \url{https://github.com/adamoshen/menace}
  \item Report bugs at \url{https://github.com/adamoshen/menace/issues}
}

}
\author{
\strong{Maintainer}: Adam Shen \email{adamshen1@gmail.com}

}
\keyword{internal}
