% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shout.R
\name{shout}
\alias{shout}
\title{Return a character vector loudly}
\usage{
shout(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
A character vector with the same length as the input.
}
\description{
Return a character vector loudly by uppercasing all letters and increasing spacing. Best used
in conjunction with \code{\link[=teams_bold]{teams_bold()}}.
}
\details{
Inspired by Ryan Kavanagh.
}
\examples{
# example code
msg <- "look here, you little shit"
shout(msg)

# Recommended to use with `teams_bold()`.
# Write to clipboard and paste to Teams
# msg |>
#   shout() |>
#   teams_bold() |>
#   clipr::write_clip()
}
